$ amount ending in 0 don't show

+++ grails-app/controllers/fedu/ProjectController.groovy
@@ -519,7 +519,7 @@ class ProjectController {
     @Secured(['ROLE_USER'])
     def customrewardsave() {
         def reward = new Reward(params)
-		int price = Integer.parseInt(params.price)
+		int price = Double.parseDouble(params.price)
 		int amount = Double.parseDouble(params.amount)
 		if(price >= amount) {
 			flash.message = "Enter a price less than Campaign amount: ${amount}"
